+++ src/interpreter/lexer/punc.js
