+++ src/interpreter/parser/ast.js
@@ -1,0 +1,12 @@
+class AST{
+    constructor(type){
+        this.type = type;
+        this.children = [];
+    }
+
+    addChild(child){
+        this.children.push(child);
+    }
+}
+
+module.exports = AST;
